backend/menu/requirements.txt
Flask==2.2.2
python-dotenv==1.0.0
google-cloud-firestore==2.7.0
google-auth==2.21.0